Correctly sends data to graphite
The requests per second and response times are sent under keys like:
    locust.POST-zones-zoneID.reqs_per_sec.host-name_b119d1c...
    locust.GET-zones-zoneID.response_time.host-name_b119d1c...
where the last portion is a locust slave id.

These should be sent to a carbon-aggregator instance (port 2023, by
default) which is set up with rules to sum or aveage over all clients
for each request. For instance, in aggregation-rules.conf:

    locust.requests.<request>.reqs_per_sec 1 = sum locust.<request>.reqs_per_sec.*
    locust.requests.<request>.response_time 5 = avg locust.<request>.response_time.*
This would sum the reqs_per_sec over all clients, but maintain separate
statistics for each request. It would average the response_times (over a
5 second interval) using data from all clients.