Web UI requires username/password authentication
Authentication is done with HTTP digest auth using a flask extension.
This is done to ensure that all Flask views will prompt for a
username/password, including those locust.web.

The config is updated to allow a single username/password combo to be
specified. If either the username or password is missing, no
authentication is setup and Locust starts as usual.