Added test persistence and a results summary page
This relies on a custom state_changed event at pglass/locust.

When the stop button is pushed, Locust's statistics are stored as json
on the master server. Opening up /reports gives a list of dated entries.
For each entry, you can open up a report (a summary page) or get the raw
json.

insight.py contains a few functions to detect which mode Locust is
running in - master, slave, or neither. These are moved out of
graphite_client.py and are used by graphite_client.py and
persistence.py.