Do data gathering during hatching
This changes data gathering to use locust tasks, so that data gathering
is rate limited instead of just spamming get requests.

* Adds the ability to switch the task set you're running, so we can
switch from data gathering tasks to actual tasks, or switch task weights
or whatever.
* Simplifies the config to be more explicit: "This tenant needs this
many zones and this many recordsets."
* Refactors ZoneInfo/RecordInfo namedtuples to a Tenant class with
simpler Zone and Recordset namedtuples. A Tenant has a list of zones
and recordsets internally, so it's more apparent which tenant owns a
zone or record.